Fixed where select site was not updateing when key file was changed. Added kyes() to the req for site_subset()
Added feature to have reference files for more then one site.  The files need to be named with all sites separated by "_".

Fixed in spectra plot when changing key files the old key file spu file name was selected and the new one had not yet been populated. Resulted in an empty file for the plot.